Split integration tests into intra-team and inter-team tests
