+++ src/integration-test/types.ts
@@ -1,3 +1,5 @@
+import CharlesClient from './charles-client';
+
 export interface Auth0 {
   domain: string;
   clientId: string;
@@ -6,17 +8,19 @@ export interface Auth0 {
 }
 export interface Config {
   charles: string;
-  notifications: {
-    flowdock?: FlowdockNotificationConfiguration,
-    hipchat?: HipChatNotificationConfiguration,
-    slack?: SlackNotificationConfiguration,
-  };
   auth0: {
     regular: Auth0 & { gitPassword: string };
     open: Auth0 & { gitPassword: string };
     admin: Auth0 & { gitPassword: string };
     [key: string]: Auth0 & { gitPassword: string };
   };
+  notifications?: NotificationConfigurations;
+}
+
+export interface NotificationConfigurations {
+  flowdock?: FlowdockNotificationConfiguration;
+  hipchat?: HipChatNotificationConfiguration;
+  slack?: SlackNotificationConfiguration;
 }
 interface HipChatNotificationConfiguration {
   type: 'hipchat';
@@ -38,4 +42,11 @@ export interface SSE {
   type: string;
   lastEventId: string;
   data: any;
+}
+
+export type TeamType = 'admin' | 'regular' | 'open';
+export interface CharlesClients {
+  admin: CharlesClient;
+  open: CharlesClient;
+  regular: CharlesClient;
 }
